Just as a suggestion, copied tutorial material structure from
https://github.com/absinthe-graphql/absinthe/tree/master/guides/tutorial

to

https://github.com/absinthe-graphql/absinthe_tutorial

with the following change suggestions

added /src subdirectory to the /tutorial directory
put the *.md files into this /src subdirectory

created two scripts (either one will do - you choose which one you like better) ,
- convert-md-to-html.sh generates the *.html files into the parent directory
- make-html-from-md.sh generates the *.md.html files into the parent directory

(the difference is in the naming of the resulting html files : *.html vs. *.md.html naming)

The resulting html files are added to the git repo, but should ideally be generated on fly from the md files.

The tool in the scripts for generating the html from md files is http://pandoc.org/MANUAL.html